Split event auth code into seperate module
